Fix Linux-Power floating point.
Noticed that Bounce and many of the dis/math applications wouldn't work.
I revisited the floating-point code and discovered I had made a real mess
of things.  Went through and cleaned stuff up and discovered the gfltconv
problems were due to me being a boofhead in other portions of the floating
point code.

This patch fixes floating point on Linux-Power and removes the gfltconv
work-around.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>