Add shared memory transport device to emu.
This device provides an ip(3) like interface to access shared memory
transports.  This patch also provides a unix-shared-memory version of
the transport.  Future extensions will include an mmap transport (which
can be used for inter-partition communication on rHype) and a xen shared
memory transport which can be used between partitions.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>