add devserve.c.
it is a bit like plan 9's srv(3).  it allows you to post open files
in the file system, so mnt(3) can mux styx on it.  as discussed on
the inferno list, there are better ideas for this.  this is a
stop-gap measure.  it's enabled in all emu builds, so it gets tested
(at least for compilation).