make garbage collector less active for emu.
execatidle() is called when there is nothing to do.  it runs the
garbage collector, which walks all memory.  that can be cpu-intensive.
it shows especially when moving the mouse over a inferno window:
the garbage collector is run after each mouse event.

this diff only runs the gc at execatidle() when the last idle gc
run was 10 seconds ago, or immediately when the memory allocation
has changed (up or down) by at least "heapmem->chunk/8" (512k/8=64k
at the moment).

this should be done for os/port/ too probably.